feat(infra): add ECR repository
- build docker imagen using nix
- push image to ECR reposiroty
