+++ .github/workflows/deploy.yml
@@ -37,3 +37,19 @@ jobs:
       - name: Deploy Frontend
         run: |
           nix run .#front-deploy
+  api-push-container:
+    needs: infra-apply
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: cachix/install-nix-action@v31
+        with:
+          nix_path: nixpkgs=channel:nixos-unstable\
+      - name: Load Age private key
+        run: |
+          mkdir -p ~/.config/sops/age
+          echo "${{ secrets.SOPS_AGE_KEY }}" > ~/.config/sops/age/keys.txt
+          chmod 600 ~/.config/sops/age/keys.txt
+      - name: Push API container
+        run: |
+          nix run .#api-container-push
