+++ shared-expenses-api/pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
     "sqlalchemy>=2.0.41,<3",
     "asyncpg>=0.30.0,<0.31",
     "psycopg2-binary>=2.9.10,<3",
+    "aioboto3>=14.3.0",
 ]
 
 [dependency-groups]
